aligo/requirements.txt
requests==2.32.2
python-dotenv==1.0.1
fastapi==0.108.0
pydantic==2.5.3
mysql-connector-python==8.4.0
uvicorn==0.29.0
python-jose
wmi
passlib